refactor(cli): move shared run helpers to run_shared.go (PRD #1009 M1)
